Collapse expression packages
