--- libs/shared/expression/evaluator/package.json
@@ -1,51 +1,0 @@
-{
-  "name": "@shipfox/expression-evaluator",
-  "license": "MIT",
-  "private": false,
-  "version": "1.0.0",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/ShipfoxHQ/shipfox.git",
-    "directory": "libs/shared/expression/evaluator"
-  },
-  "type": "module",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "imports": {
-    "#*": "./src/*"
-  },
-  "files": [
-    "dist"
-  ],
-  "exports": {
-    ".": {
-      "development": {
-        "types": "./src/index.ts",
-        "default": "./src/index.ts"
-      },
-      "default": {
-        "types": "./dist/index.d.ts",
-        "default": "./dist/index.js"
-      }
-    }
-  },
-  "scripts": {
-    "build": "shipfox-swc",
-    "check": "shipfox-biome-check",
-    "check:fix": "shipfox-biome-check --write",
-    "test": "shipfox-vitest-run",
-    "type": "shipfox-tsc-check",
-    "type:emit": "shipfox-tsc-emit"
-  },
-  "dependencies": {
-    "@gresb/cel-javascript": "0.1.1",
-    "@shipfox/expression-language": "workspace:*"
-  },
-  "devDependencies": {
-    "@shipfox/biome": "workspace:*",
-    "@shipfox/swc": "workspace:*",
-    "@shipfox/ts-config": "workspace:*",
-    "@shipfox/typescript": "workspace:*",
-    "@shipfox/vitest": "workspace:*"
-  }
-}
