--- libs/shared/expression/evaluator/src/index.ts
@@ -1,8 +1,0 @@
-export {
-  evaluateWorkflowExpression,
-  evaluateWorkflowPredicate,
-  type WorkflowExpressionEvaluationContext,
-  WorkflowExpressionEvaluationError,
-  type WorkflowExpressionEvaluationValue,
-  workflowExpressionEvaluationErrorCode,
-} from '#evaluator/index.js';
