--- libs/shared/expression/language/src/index.ts
@@ -1,14 +1,0 @@
-export {
-  type CreateWorkflowExpressionParams,
-  createWorkflowExpression,
-  type ExpressionScalarType,
-  type ExpressionType,
-  type ExpressionTypeEnvironment,
-  InvalidWorkflowExpressionError,
-  invalidWorkflowExpressionErrorCode,
-  unsafeWorkflowExpressionFromSource,
-  type ValidCelExpression,
-  type WorkflowExpression,
-  type WorkflowExpressionCheck,
-  type WorkflowExpressionCheckOptions,
-} from './expression/index.js';
