--- libs/shared/expression/language/tsconfig.build.json
@@ -1,9 +1,0 @@
-{
-  "extends": "@shipfox/ts-config/node",
-  "compilerOptions": {
-    "rootDir": "src",
-    "outDir": "dist"
-  },
-  "include": ["src"],
-  "exclude": ["**/*.test.tsx", "**/*.test.ts"]
-}
