--- libs/shared/expression/language/tsconfig.json
@@ -1,3 +1,0 @@
-{
-  "references": [{ "path": "tsconfig.build.json" }, { "path": "tsconfig.test.json" }]
-}
