--- libs/shared/expression/language/vitest.config.ts
@@ -1,3 +1,0 @@
-import {defineConfig, type UserConfigExport} from '@shipfox/vitest';
-
-export default defineConfig({}, import.meta.url) as UserConfigExport;
