+++ libs/shared/expression/src/index.ts
@@ -1,0 +1,28 @@
+export {
+  WorkflowExpressionEvaluationError,
+  workflowExpressionEvaluationErrorCode,
+} from './evaluator/errors.js';
+export {
+  evaluateWorkflowExpression,
+  evaluateWorkflowPredicate,
+  type WorkflowExpressionEvaluationContext,
+  type WorkflowExpressionEvaluationValue,
+} from './evaluator/evaluate-workflow-expression.js';
+export {
+  createWorkflowExpression,
+  unsafeWorkflowExpressionFromSource,
+} from './expression/create-workflow-expression.js';
+export {
+  InvalidWorkflowExpressionError,
+  invalidWorkflowExpressionErrorCode,
+} from './expression/errors.js';
+export type {
+  CreateWorkflowExpressionParams,
+  ExpressionScalarType,
+  ExpressionType,
+  ExpressionTypeEnvironment,
+  ValidCelExpression,
+  WorkflowExpression,
+  WorkflowExpressionCheck,
+  WorkflowExpressionCheckOptions,
+} from './expression/workflow-expression.js';
